_scripts/requirements.txt
pyquery==1.2.4
PyYAML==3.10